Initial implementation of C++ password checker
This tool uses Boost Asio and multiple threads to max out network and
CPU usage. It also uses Boost's coroutines for slightly more readable
code. SHA1 calculations are done with OpenSSL's implementation.

The tool has the option to specify a timeout to be applied to requests,
in milliseconds, and whether to use a cache or not (which is never
persisted anyway). Using a cache can put the bottleneck on
multithreaded synchronisation instead of networking, so *not* using it
could be worth in some situations.

The tool does support a "dry-run" mode where no network communication
happens. This was useful to test the rest of the logic, but is also
useful to show what the maximum limit is for checks per second on your
machine.

For convenience there is a Dockerfile to build a Docker image with a
Release build of the tool installed. This is based on Ubuntu 21.10, and
adds only a couple of megabytes on top of the original image, making it
fairly lightweight.